+++ tests/ApiTest.php
@@ -24,16 +24,16 @@ it('can convert 27823096710 to +27.82 309 6710', function () {
     expect($result)->toEqual('+27.82 309 6710');
 });
 
-it('can post an array to a slack', function () {
+it('can post an Array to a Slack', function () {
     require 'vendor/autoload.php';
 
     $dotenv = Dotenv\Dotenv::createImmutable(__DIR__ . "/..");
     $dotenv->load();
 
     $data = [
-        "token" => $_ENV['SLACK_TEST_TOKEN'],
-        "channel" => $_ENV['SLACK_TEST_CHANNEL'],
-        "text" => "Hello, Foo-Bar channel message.",
+        "token"    => $_ENV['SLACK_TEST_TOKEN'],
+        "channel"  => $_ENV['SLACK_TEST_CHANNEL'],
+        "text"     => "Test 1",
         "username" => "tester",
     ];
 
@@ -46,3 +46,31 @@ it('can post an array to a slack', function () {
 
     expect(json_decode($result, true)["ok"])->toEqual(true);
 });
+
+it('can post a String to a Slack', function () {
+    require 'vendor/autoload.php';
+
+    $dotenv = Dotenv\Dotenv::createImmutable(__DIR__ . "/..");
+    $dotenv->load();
+
+    $header = [
+        "Content-Type: application/json;charset=utf-8",
+        "Authorization: Bearer {$_ENV['SLACK_TEST_TOKEN']}",
+    ];
+
+    $data = "{'channel': '" . $_ENV['SLACK_TEST_CHANNEL'] . "',"
+            . "'text': 'Test 2',"
+            . "'username': 'tester'}";
+    
+    $api = new Api();
+
+    $result = $api->post(
+        'https://slack.com/api/chat.postMessage',
+        $data,
+        $header
+    );
+    
+    expect(json_decode($result, true)["ok"])->toEqual(true);
+});
+
+